pyproject.toml
[tool.poetry]
name = "hexlet-flask-example"
version = "0.1.0"
description = ""
authors = ["Asgef <asgefes1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
flask-wtf = "^1.2.1"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
